requirements.txt
certifi==2019.6.16
chardet==3.0.4
clearbit==0.1.7
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
email-hunter-python==1.1.0
Faker==2.0.1
idna==2.8
pkg-resources==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
